add birthday parameter on login user(for updating old users) redirect to edit user profile after registration fix some leak issus add caption keyboard disappears on Done.
git-svn-id: svn://mobilep.ro/status@242 99022451-f8c5-4652-be14-efb92feb6bba

+++ trunk/Status/ViewControllers/Chat/cells/STConversationCell.m
@@ -96,7 +96,7 @@ static NSString *kOnlineImageName = @"online chat";
     
     NSInteger numberOfUnreadMessages = [info[@"unread_messages_count"] integerValue];
     if (numberOfUnreadMessages > 0) {
-        [_numberOfUnreadMessages setTitle:[NSString stringWithFormat:@"%ld", numberOfUnreadMessages] forState:UIControlStateNormal];
+        [_numberOfUnreadMessages setTitle:[NSString stringWithFormat:@"%ld", (long)numberOfUnreadMessages] forState:UIControlStateNormal];
         _numberOfUnreadMessages.hidden = NO;
     }
     else
